backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-JWT-Extended
PyJWT
requests
pillow
opencv-python
facenet-pytorch
scikit-learn
numpy
torch
torchvision 